+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- support for configuring the SSH connection timeout, defaults to 10 seconds
+
 ## 0.8.4 - 2026-01-20
 
 ### Fixed